Only set the current active row to output and the unused rows to input, which essentially makes them HiZ for output purposes.
If the unused rows actively output a "0" (gnd) then if two keys on the same column are pressed, the key on the inactive row will pull the active row down and no presses are registered.
Also if one of the keys has already registered a "pressed" transition, it will then register a "released" transition even if it hadn't been released.
Floating unused outputs prevents this cross-row pulldown from happening.